Enable directory mounting for nginx and gunicorn
