+++ Vagrantfile
@@ -4,6 +4,7 @@
 Vagrant.configure("2") do |config|
   config.vm.box = "bento/ubuntu-16.04"
 
+  config.vm.network "forwarded_port", guest: 80, host: 8080, id: "nginx"
   config.vm.network "forwarded_port", guest: 5000, host: 5000, id: "Thumbtack-dev"
   config.vm.network "forwarded_port", guest: 8208, host: 8208, id: "Thumbtack"
   config.vm.provider "virtualbox" do |v|
