+++ src/thumbtack/__init__.py
@@ -1,7 +1,5 @@
 import logging
 import os
-import time
-import threading
 
 from pathlib import Path
 from pkg_resources import get_distribution, DistributionNotFound
@@ -11,6 +9,7 @@ import click
 from flask import Flask, current_app
 from flask_restful import Api
 
+from .directory_monitoring import DirectoryMonitoring
 from .resources import Mount, SupportedLibraries
 from .utils import init_db, monitor_image_dir
 from .views import main
@@ -20,18 +19,6 @@ try:
     __version__ = get_distribution('thumbtack').version
 except DistributionNotFound:
     __version__ = 'Could not find version'
-
-
-class directory_monitoring(threading.Thread):
-    def __init__(self, app):
-        threading.Thread.__init__(self)
-        self.app = app
-
-    def run(self):
-        with self.app.app_context():
-            while True:
-                time.sleep(3)
-                monitor_image_dir()
 
 
 def create_app(image_dir=None, database=None):
@@ -104,16 +91,17 @@ def before_first_request():
 
 
 def configure_logging(app):
-    if not app.debug:
-        # In production mode, add log handler to sys.stderr.
+    formatter = logging.Formatter("[%(asctime)s] %(levelname)s in %(name)s.%(module)s: %(message)s")
+
+    if app.debug:
         app.logger.setLevel(logging.DEBUG)
+    else:
+        app.logger.setLevel(logging.INFO)
 
-        formatter = logging.Formatter("[%(asctime)s] %(levelname)s in %(name)s.%(module)s: %(message)s")
-
+        # In production mode, add log handler to sys.stderr.
         shandler = logging.StreamHandler()
-        shandler.setLevel(logging.DEBUG)
+        shandler.setLevel(logging.INFO)
         shandler.setFormatter(formatter)
-
         # app.logger.addHandler(shandler)
 
 
@@ -128,6 +116,6 @@ def configure_logging(app):
 @click.option('--db', 'database', help='SQLite database to store mount state')
 def start_app(debug, host, port, image_dir, database):
     app = create_app(image_dir=image_dir, database=database)
-    directory_monitoring_thread = directory_monitoring(app)
+    directory_monitoring_thread = DirectoryMonitoring(app)
     directory_monitoring_thread.start()
     app.run(debug=debug, host=host, port=port)
