+++ src/thumbtack/views.py
@@ -23,9 +23,11 @@ def index():
         else:
             unsupported_mount_types.append(mount_type)
 
+    current_app.logger.debug('-------------- Getting images!!! --------------')
     # On refresh, update DB to match what's on disk
     monitor_image_dir()
     images = get_images()
+    current_app.logger.debug('-------------- Got images!!! --------------')
 
     return render_template('index.html',
                            supported_mount_types=supported_mount_types,
